fix: Fix broken Mocha tests.
This fixes rendering for connection tests, switches out a now-contrived
output connection test, and revises FocusManager state such that a
unique FocusManager is created for every individual test to avoid
cross-test state bleeding. Focus state can still bleed, unfortunately,
but FocusManager generally guarantees eventual consistency.
